PWN-9267, PWN-9352 - Striga. Integrated withdrawal + signup flow to striga off-ramp calculator (#2008)
- fixed UiKitEditText state saving
- fixed striga withdraw "receiver" text appearance
- fixed text appearance for striga swap calculator
- fixed striga calculator progress button
- fixed disabling exchange rate polling when error occurred
- fixed striga statement exchangeRate value from int to BigDecimal
- fixed inputTextColor on NewUiKitEditText
- changed popBackStack to popBackStackTo on StrigaOnboardingFragment

+++ app/src/main/java/org/p2p/wallet/striga/wallet/api/response/StrigaAccountStatementResponse.kt
@@ -1,6 +1,7 @@
 package org.p2p.wallet.striga.wallet.api.response
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 data class StrigaAccountStatementResponse(
     @SerializedName("transactions")
@@ -28,7 +29,7 @@ data class StrigaAccountStatementResponse(
         @SerializedName("feeEur")
         val feeEur: Long,
         @SerializedName("exchangeRate")
-        val exchangeRate: Long,
+        val exchangeRate: BigDecimal,
         @SerializedName("balanceBefore")
         val balanceBefore: AccountBalanceResponse,
         @SerializedName("balanceAfter")
